requirements.txt
Flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
gspread==5.12.0
google-auth==2.25.2
python-dotenv==1.0.0
Werkzeug==3.0.1
